breakdown structs file as its too large
